Remove redundant BaseMongoConnectorTest
